requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-import-export==3.3.3
et-xmlfile==1.1.0
fonttools==4.45.0
freetype-py==2.3.0
html5lib==1.1
idna==3.4
lxml==4.9.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.17.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
weasyprint==60.1
webencodings==0.5.1
xhtml2pdf==0.2.13
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
